move blog components
